+++ openhands/integrations/github/service/__init__.py
@@ -2,16 +2,15 @@
 
 from .branches_prs import GitHubBranchesMixin
 from .features import GitHubFeaturesMixin
-from .microagents import GitHubMicroagentsMixin
 from .prs import GitHubPRsMixin
 from .repos import GitHubReposMixin
 from .resolver import GitHubResolverMixin
 
 __all__ = [
-    "GitHubBranchesMixin",
-    "GitHubFeaturesMixin",
-    "GitHubMicroagentsMixin",
-    "GitHubPRsMixin",
-    "GitHubReposMixin",
-    "GitHubResolverMixin",
+    'GitHubBranchesMixin',
+    'GitHubFeaturesMixin',
+    'GitHubMicroagentsMixin',
+    'GitHubPRsMixin',
+    'GitHubReposMixin',
+    'GitHubResolverMixin',
 ]
