+++ openhands/integrations/github/service/repos.py
@@ -7,6 +7,10 @@ from openhands.server.types import AppMode
 
 
 class GitHubReposMixin(GitHubMixinBase):
+    """
+    Methods for interacting with GitHub repositories (from both personal and app installations)
+    """
+
     async def get_installations(self) -> list[str]:
         url = f'{self.BASE_URL}/user/installations'
         response, _ = await self._make_request(url)
